pyproject.toml
[project]
name = "diamond-setup"
version = "1.0.0"
description = "Universal Python project scaffold — generate professional project skeletons in seconds"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "GenesisAeon" }]
requires-python = ">=3.11"
keywords = ["scaffold", "template", "cli", "project-generator", "skeleton"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
    "pydantic>=2.7",
]

[project.optional-dependencies]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]
dev = [
    "ruff>=0.6.0",
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
diamond = "diamond_setup.cli:app"

[project.urls]
Repository = "https://github.com/GenesisAeon/diamond-setup"
Issues = "https://github.com/GenesisAeon/diamond-setup/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/diamond_setup"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "W", "UP", "C4", "SIM"]
ignore = ["B008"]  # typer uses function calls in defaults

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # allow assert in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=diamond_setup --cov-report=term-missing -v"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
